backend/requirements.txt
google-adk>=1.0.0
google-genai>=1.0.0
chromadb>=0.5.0
google-search-results>=2.4.2
streamlit>=1.38.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.0
